% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binom_fit_lazy}
\alias{binom_fit_lazy}
\title{Fit Logistic Regression with AdaGrad SGD}
\usage{
binom_fit_lazy(X, y, m = 1L, lambda = 0, minibatch = 1L,
  max_epochs = 1000L, step_scale = 1, tol = 1e-06, conv_autocor = 0.06,
  history = FALSE, eval_every = 1L)
}
\arguments{
\item{beta:}{numeric vector of parameters of linear model}

\item{X:}{numeric matrix with predictor variables}

\item{y:}{numeric vector with response variable}

\item{m:}{integer controlling the max number of successes in binomial response y}
}
\value{
matrix with one column with the gradient of the negative loglikelihood evaluated at the input
}
