\name{fastsgd-package}
\alias{fastsgd-package}
\alias{fastsgd}
\docType{package}
\title{
\packageTitle{fastsgd}
}
\description{
\packageDescription{fastsgd}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fastsgd}
\packageIndices{fastsgd}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{fastsgd}

Maintainer: \packageMaintainer{fastsgd}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
