% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nlogl_binom}
\alias{nlogl_binom}
\title{Negative LogLikelihood of Binomial GLM}
\usage{
nlogl_binom(beta, X, y, m)
}
\arguments{
\item{beta:}{numeric vector of parameters of linear model}

\item{X:}{numeric matrix with predictor variables}

\item{y:}{numeric vector with response variable}

\item{m:}{integer controlling the max number of successes in binomial response y}
}
\value{
the negative loglikelihood evaluated at the input
}
