% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nlogl_binom_grad}
\alias{nlogl_binom_grad}
\title{Gradient Negative LogLikelihood of Binomial GLM}
\usage{
nlogl_binom_grad(beta, X, y, m = 1L, lambda = 0)
}
\arguments{
\item{beta:}{numeric vector of parameters of linear model}

\item{X:}{numeric matrix with predictor variables}

\item{y:}{numeric vector with response variable}

\item{m:}{integer controlling the max number of successes in binomial response y}
}
\value{
matrix with one column with the gradient of the negative loglikelihood evaluated at the input
}
