% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid function with truncation}
\usage{
sigmoid(A)
}
\arguments{
\item{u:}{numeric vector of input}

\item{eps:}{real number of truncation paramaeter if output is close to 0/1}
}
\value{
sigmoid evaluate at input u with truncation
}
