% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sparse_logit}
\alias{sparse_logit}
\title{Sparse logistic with lasso penalizer}
\usage{
sparse_logit(beta0, Xt, y, lambda = 0, maxepochs = 1L, step_scale = 2,
  tol = 0.001, discount = 0.95)
}
\description{
Fits a sparse logistic model with lasso penalizer
}
